pyproject.toml
[project]
name = "patch-camelyon"
version = "0.1.0"
authors = [{ name = "Matěj Pekár", email = "matejpekar@mail.muni.cz" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "torch>=2.4.0",
    "lightning>=2.3.3",
    "tqdm>=4.66.4",
    "albumentations>=1.4.13",
    "hydra-core>=1.3.2",
    "torchvision>=0.19.0",
    "torchmetrics>=1.4.1",
    "rationai-mlkit @ git+https://gitlab.ics.muni.cz/rationai/digital-pathology/libraries/mlkit.git",
]

[project.optional-dependencies]
preprocessing = ["h5py>=3.11.0"]

[tool.pdm.dev-dependencies]
dev = ["mypy>=1.11.2", "pre-commit>=3.8.0", "ruff>=0.6.4"]


[tool.pdm.scripts]
fit = "python -m patch_camelyon mode=fit"
validate = "python -m patch_camelyon mode=validate"
test = "python -m patch_camelyon mode=test"
predict = "python -m patch_camelyon mode=predict"
l = { composite = ["lint", "format", "mypy"] }
lint = "ruff check --fix"
format = "ruff format"
mypy = "mypy ."
post_install = { composite = [
    "pre-commit autoupdate",
    "pre-commit install",
    "pre-commit install --hook-type commit-msg",
] }

[tool.pdm]
distribution = false
